mavlink: move CAMERA_IMAGE_CAPTURED to separate stream header
